move app component to components directory
