+++ src/components/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Provider } from 'react-redux'
-import store from './redux/store'
-import chooseWinner from './redux/actionCreators'
-import Raffle from './containers/RaffleContainer'
+import store from '../redux/store'
+import chooseWinner from '../redux/actionCreators'
+import Raffle from '../containers/RaffleContainer'
 
 const App = React.createClass({
   handleClick() {
